Change CI action image to ubuntu-16.04
This may be a problem with the [openssl version](https://stackoverflow.com/questions/54123750/why-is-herokus-heroku-18-stack-only-compatible-with-puma-versions-3-7-x) due to commonalities in the error message.

If so, we can either version down openssl -- which versioning down ubuntu to 16.04 should get us -- or we can version up puma to 3.7. The latter will need to be done eventually but this might do for a quick fix