Rework Build and Test action with Matt Swanson's template
Perhaps needless to say, there's no magic here doing the db setup for us, so that was the next error we ran into.

There's several ways to go around this, but I instantly started reaching for other projects where this is a solved problem.

Found Matt Swanson's template and am largely running with that.

I noticed there's no ruby version spec'ed in it so we went ahead and set that.

Not setting it in a `.ruby-version` file because we don't need to make that decision yet, and right this second I want the project standard for non-dockerized development to be asdf (assuming we intend to support non-dockerized development).