Add .tool-versions
Well we ran into this on the setup ruby step:
`Error: input ruby-version needs to be specified if no .ruby-version or .tool-versions file exists`

Spec-ing the ruby version in the workflow would be a less-invasive and less far-reaching change. I may regret it, but since my heart is already set on asdf, I'm adding a .tool-versions.

Going ahead and spec-ing the node version there as well and optimistically removing the spec from the pertinent node build steps.