Add postgres healthcheck, change postgres config, change some formatting
The main things we're doing to the CI action here is adding a healthcheck for postgres when we set up the container, changing the `PGHOST` env variable to the postgres container, and changing . We were getting bad connection errors within the CI action container telling us to make sure there's an accessible service running on port 5432.

It's strange that this couldn't at least get to a further-along db error out-of-the-box. Matt's blog post might be out of date.

The simple way to confirm this would be to only change the POSTGRES_DB here and confirm we get the same error, but based on the error I'm pretty confident we're dealing with pointing at the wrong host:

```
could not connect to server: Connection refused
	Is the server running on host "localhost" (::1) and accepting
	TCP/IP connections on port 5432?
```