Add project directory and first work items
This is experimental, and part of my dream of putting project management in source control -- really and truly, everything in source control that is reasonable to have there.

Note that for starters and depending on how you look at it, there is either no notion of a work item ID, or the de facto work item ID (the filename) is coupled to the work item title.

There is little reason to believe this would be feasible in the long term, but as it makes the current work item-modifying workflow easier to deal with and I'm not putting a great deal of thought presently into what interface should sit between a user and the data representations of these work items, it seems good for now.