Change PGHOST to 127.0.0.1
We know `localhost` doesn't do the trick, and we ran into trouble with the working directory we ended up in to make `postgres` work.

Another simple angle to come at this from would be to keep the PGHOST of postgres and then bundle exec our rails commands though I suspect we'd run into similar issues.