More tweaks
Only thing I'll do on this today.

We're moving the env definitions to the top of the tests job, which puts all the Postgres-y configgy bits together. I saw this in the GitHub job for papercups over in Elixirland.

Over in the `database.yml, we're making sure to actually grab the values set in the env for our tests job.

Ultimately I don't want defaults here. The app depends on a secure db, and to get at it it needs credentials. These should neither be optional nor their source hard to decipher.

For now we are allowing defaults for username and password to avoid a willfull breaking deployment change for production.

Though, it must be admitted, we don't actually know if this version of the app is working with the updated deps so...🤷

Actually now that I see it, I feel really silly because when I look at the early error messages for this I feel like it was screaming in the face at me that I didn't give it the right stuff: `Couldn't create database for {"adapter"=>"postgresql", "encoding"=>"unicode", "pool"=>5, "database"=>"muscle_wizards_test"}`. It's the following "Is the service up and running and the port accepting connections?" messages that put me off track.

Notably, the postgres health check was failing, but I also think it may have been exclusively the super-speedy ones. I've slowed it back down to give it a chance to go green. If it fails anyway I won't feel as embarassed.