Revert "Merge pull request #85 from SherSpock/version-all-the-things-or-at-least-some-of-them"
This broke stuff. That was fine and in-line to merge given the philosophy of the PR of "break it now and sort it out later". We also tagged the last good state.

Since then, I've decided that the work I have planned for MuscleWizards -- which is really more of a learning project for me than anything else, although if completed it likely will have considerable utility nonetheless -- ought not to continue in this repo.

AFAIK, this repo is still the canonical MuscleWizards repository, and pushing forward with a range of changes without waiting for input on a project created in partnership would be rude.